Move services into /services
